Initial Commit
TODOS: Add ability to handle an odd number of players.  Person with the
worst record gets the bye for each round unless that person has already
had a bye.  In that case, then the next worst player will receive the
bye.  A bye counts as a win and will be stored in the database as NULL.
 Conditionals will check for NULL when computation/aggregation is
required (e.g. calculating O.M.W.).