Odd players; README.md
Added ability to handle an odd number of players.  Person with the
worst record gets the bye for each round unless that person already had
a bye.  In that case, the next worst player will receive the bye.  A
bye counts as a win and will be stored in the match_history table as -1.
Conditionals check for -1 when computation/aggregation is required
(e.g. calculating O.M.W.).