Removed finalStandings() method; Refactored playerStandings(); player_standings VIEW implemented
Refactored the player standings method with the PostgreSQL View
‘player_standings’ to calculate the standings on the server side.
Added another UNION statement to the ‘player_standings’ view to handle
by matches and to ensure that players with bye matches are not omitted.

If the simulation is in its first round, then the match_history table
will be empty and the function will manually query the DB to populate
the standings for pairing purposes.